requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Django==4.1.12
django-admin-argon-dashboard==1.0.16
django-admin-soft-dashboard==1.0.12
dnspython==2.4.2
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
numpy==1.26.2
packaging==23.2
pymongo==4.6.0
python-decouple==3.4
python-dotenv==1.0.0
pytz==2023.3.post1
rpds-py==0.13.0
scipy==1.11.4
six==1.16.0
sqlparse==0.2.4
typing_extensions==4.8.0
tzdata==2023.3
vboxapi==1.0
Werkzeug==3.0.1
whitenoise==6.5.0
WTForms==2.3.3

djangorestframework~=3.14.0